backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
torch>=2.1.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
pillow>=10.0.0
sentencepiece>=0.2.0
imageio>=2.34.0
imageio-ffmpeg>=0.5.0
